Move default views and Direction to own file
