+++ pytissueoptics/rayscattering/views/__init__.py
@@ -1,4 +1,3 @@
-from .view2D import Direction, ViewGroup, View2D, \
-    View2DProjection, View2DProjectionX, View2DProjectionY, View2DProjectionZ, \
-    View2DSurface, View2DSurfaceX, View2DSurfaceY, View2DSurfaceZ
+from .view2D import ViewGroup, View2D
+from .direction import Direction
 from .viewFactory import ViewFactory
